findandfriendserver/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
et-xmlfile==1.1.0
Flask==3.0.3
flask-mysql-connector==1.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==2.1.5
MouseInfo==0.1.3
mysql-connector-python==9.0.0
ninja==1.11.1.1
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
pillow==10.3.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyJWT==2.9.0
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.6
zipp==3.20.2
